refactored folder structure in book details folder
