+++ src/app/books/[olid]/components/Overview/Overview.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import IBook from '../../Book.types'
-import { Flex } from '@radix-ui/themes'
+import IBook from '../../../Book.types'
+import { Box, Flex, Text } from '@radix-ui/themes'
 import TopSection from './TopSection'
 import Cards from './Cards'
 
@@ -9,6 +9,14 @@ export default function Overview({ work, book }: { work: any; book: IBook }) {
     <>
       <Flex direction='column' gap={{ initial: '3', md: '4' }}>
         <TopSection work={work} book={book} />
+        <Box data-testid='book-details-first-sentence'>
+          <Text size='4' weight='regular'>
+            First Sentence:{' '}
+            <Text className='italic' weight='light'>
+              {work.first_sentence}
+            </Text>
+          </Text>
+        </Box>
         <Cards work={work} book={book} />
       </Flex>
     </>
